add sampler and assigner registry
