+++ mmdet/core/anchor/guided_anchor_target.py
@@ -1,6 +1,7 @@
 import torch
 
-from ..bbox import PseudoSampler, build_assigner, build_sampler
+from ..bbox import build_assigner, build_sampler
+from ..bbox.samplers.pseudo_sampler import PseudoSampler
 from ..utils import multi_apply, unmap
