+++ mmdet/core/bbox/__init__.py
@@ -1,22 +1,17 @@
-from .assigners import AssignResult, BaseAssigner, MaxIoUAssigner
+from .assigners import AssignResult
 from .bbox_target import bbox_target
 from .geometry import bbox_overlaps
-from .samplers import (BaseSampler, CombinedSampler,
-                       InstanceBalancedPosSampler, IoUBalancedNegSampler,
-                       PseudoSampler, RandomSampler, SamplingResult)
+from .samplers import SamplingResult
 from .transforms import (bbox2delta, bbox2result, bbox2roi, bbox_flip,
                          bbox_mapping, bbox_mapping_back, delta2bbox,
                          distance2bbox, roi2bbox)
 
-from .assign_sampling import (  # isort:skip, avoid recursive imports
+from .builder import (  # isort:skip, avoid recursive imports
     assign_and_sample, build_assigner, build_sampler)
 
 __all__ = [
-    'bbox_overlaps', 'BaseAssigner', 'MaxIoUAssigner', 'AssignResult',
-    'BaseSampler', 'PseudoSampler', 'RandomSampler',
-    'InstanceBalancedPosSampler', 'IoUBalancedNegSampler', 'CombinedSampler',
-    'SamplingResult', 'build_assigner', 'build_sampler', 'assign_and_sample',
-    'bbox2delta', 'delta2bbox', 'bbox_flip', 'bbox_mapping',
-    'bbox_mapping_back', 'bbox2roi', 'roi2bbox', 'bbox2result',
-    'distance2bbox', 'bbox_target'
+    'bbox_overlaps', 'AssignResult', 'SamplingResult', 'build_assigner',
+    'build_sampler', 'assign_and_sample', 'bbox2delta', 'delta2bbox',
+    'bbox_flip', 'bbox_mapping', 'bbox_mapping_back', 'bbox2roi', 'roi2bbox',
+    'bbox2result', 'distance2bbox', 'bbox_target'
 ]
