+++ mmdet/core/bbox/assigners/approx_max_iou_assigner.py
@@ -1,9 +1,11 @@
 import torch
 
 from ..geometry import bbox_overlaps
+from ..registry import ASSIGNERS
 from .max_iou_assigner import MaxIoUAssigner
 
 
+@ASSIGNERS.register_module
 class ApproxMaxIoUAssigner(MaxIoUAssigner):
     """Assign a corresponding gt bbox or background to each bbox.
