+++ mmdet/core/bbox/assigners/atss_assigner.py
@@ -1,10 +1,12 @@
 import torch
 
 from ..geometry import bbox_overlaps
+from ..registry import ASSIGNERS
 from .assign_result import AssignResult
 from .base_assigner import BaseAssigner
 
 
+@ASSIGNERS.register_module
 class ATSSAssigner(BaseAssigner):
     """Assign a corresponding gt bbox or background to each bbox.
