+++ mmdet/core/bbox/assigners/point_assigner.py
@@ -1,9 +1,11 @@
 import torch
 
+from ..registry import ASSIGNERS
 from .assign_result import AssignResult
 from .base_assigner import BaseAssigner
 
 
+@ASSIGNERS.register_module
 class PointAssigner(BaseAssigner):
     """Assign a corresponding gt bbox or background to each point.
