+++ mmdet/core/bbox/registry.py
@@ -1,0 +1,4 @@
+from mmdet.utils import Registry
+
+ASSIGNERS = Registry('assigner')
+SAMPLERS = Registry('sampler')
