+++ mmdet/core/bbox/samplers/instance_balanced_pos_sampler.py
@@ -1,9 +1,11 @@
 import numpy as np
 import torch
 
+from ..registry import SAMPLERS
 from .random_sampler import RandomSampler
 
 
+@SAMPLERS.register_module
 class InstanceBalancedPosSampler(RandomSampler):
 
     def _sample_pos(self, assign_result, num_expected, **kwargs):
