+++ mmdet/core/bbox/samplers/ohem_sampler.py
@@ -1,9 +1,11 @@
 import torch
 
+from ..registry import SAMPLERS
 from ..transforms import bbox2roi
 from .base_sampler import BaseSampler
 
 
+@SAMPLERS.register_module
 class OHEMSampler(BaseSampler):
     """
     Online Hard Example Mining Sampler described in [1]_.
