+++ mmdet/core/bbox/samplers/pseudo_sampler.py
@@ -1,9 +1,11 @@
 import torch
 
+from ..registry import SAMPLERS
 from .base_sampler import BaseSampler
 from .sampling_result import SamplingResult
 
 
+@SAMPLERS.register_module
 class PseudoSampler(BaseSampler):
 
     def __init__(self, **kwargs):
