+++ mmdet/core/bbox/samplers/random_sampler.py
@@ -1,8 +1,10 @@
 import torch
 
+from ..registry import SAMPLERS
 from .base_sampler import BaseSampler
 
 
+@SAMPLERS.register_module
 class RandomSampler(BaseSampler):
 
     def __init__(self,
